Add a copy for Galaxy.csv
Add a copy of Galaxy.csv named Galaxy_copy.csv so we can practice editing one of them to our heart's content without messing up the original data. I know that Erin has her own copy, so we wouldn't be deleting data from the lab, but I still think it could be helpful to have an unedited copy to reference.