Separate Files
I put all the classes and functions into separate files. I also finished the searching functionality so that a user can pull up the data in one column for each galaxy.

I attempted to implement the add/update function to no avail. One crucial part is being able to write back to the csv file so that it truly updates the data but I could not for the life of me figure it out without the program deleting the entire csv file 😭. The code is there for the add/update function, we just need to figure out how to properly write back to the csv file.